Refactors inventory-related logic
