+++ types/items.js
@@ -11,16 +11,3 @@
 * @prop {string} descriptionFlavor 
 * @prop {boolean} lightSource   
 */
-
-/**
-* @typedef InventoryGUI 
-* @prop {Array<Item>} inventory  
-* @prop {Array<Array<ItemDescriptionToggle>>} itemRowToggle 
-* @prop {Array<ItemDescriptionToggle>} itemDescriptionToggle 
-*/
-
-/** 
-* @typedef ItemDescriptionToggle 
-* @prop {string} className  
-* @prop {string} target 
-*/
